+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Change Log
 
 All notable changes to the "dependobuf" extension will be documented in this file.
-
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## Unreleased
 
-- Initial release
+### Added
+
+- Basic DependoBuf extension.
+- Context action to install `dbuf-lsp`.